% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{mixed_covariance}
\alias{mixed_covariance}
\title{Specify a covariance structure}
\usage{
mixed_covariance(
  type = c("ar1", "car1", "cs", "arma", "toep", "us", "diag", "ou", "exp", "gau", "mat"),
  time = NULL,
  group = NULL,
  spatial = NULL
)
}
\arguments{
\item{type}{Covariance structure.}

\item{time}{Time or ordered coordinate variable.}

\item{group}{Grouping variable.}

\item{spatial}{Optional two-dimensional coordinate names.}
}
\value{
A \code{mixedflow_covariance} specification.
}
\description{
Specify a covariance structure
}
\examples{
# Example 1: discrete AR(1)
mixed_covariance("ar1", "time", "subject")
# Example 2: continuous AR(1)
mixed_covariance("car1", "day", "plant")
# Example 3: Matern spatial structure
mixed_covariance("mat", group="field", spatial=c("row","col"))
}
