% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss.R
\name{mixed_gamlss_quantiles}
\alias{mixed_gamlss_quantiles}
\title{Obtain fitted conditional quantiles from a GAMLSS model}
\usage{
mixed_gamlss_quantiles(object, probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{object}{Fitted GAMLSS object.}

\item{probs}{Probabilities.}
}
\value{
Long data frame with fitted quantiles for observed rows.
}
\description{
Obtain fitted conditional quantiles from a GAMLSS model
}
\examples{
d<-mixed_data("longitudinal")
# Example 1: quartiles
\donttest{
  if (requireNamespace("gamlss", quietly = TRUE) && requireNamespace("gamlss.dist", 
      quietly = TRUE)) {
      m <- mixed_gamlss(height ~ time, d, random = "subject")
      mixed_gamlss_quantiles(m, c(0.25, 0.5, 0.75))
  }
}
# Example 2: median
\donttest{
  if (requireNamespace("gamlss", quietly = TRUE) && requireNamespace("gamlss.dist", 
      quietly = TRUE)) {
      m <- mixed_gamlss(height ~ time, d, random = "subject")
      mixed_gamlss_quantiles(m, 0.5)
  }
}
# Example 3: central 95 percent limits
\donttest{
  if (requireNamespace("gamlss", quietly = TRUE) && requireNamespace("gamlss.dist", 
      quietly = TRUE)) {
      m <- mixed_gamlss(height ~ time, d, sigma.formula = ~treatment, random = "subject")
      mixed_gamlss_quantiles(m, c(0.025, 0.975))
  }
}
}
