% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{mixed_pp_check}
\alias{mixed_pp_check}
\title{Posterior predictive check}
\usage{
mixed_pp_check(object, type = "dens_overlay", ndraws = 100, ...)
}
\arguments{
\item{object}{Bayesian fit.}

\item{type}{\code{bayesplot}/\code{brms} predictive-check type.}

\item{ndraws}{Number of predictive draws displayed.}

\item{...}{Additional \code{brms::pp_check} arguments.}
}
\value{
A posterior predictive check plot.
}
\description{
Posterior predictive check
}
\examples{
# Example 1: density overlay
\dontrun{
  d <- mixed_data("longitudinal")
  m <- mixed_bayes(height ~ time + (1 | subject), d, chains = 2, iter = 500)
  mixed_pp_check(m)
}
# Example 2: ECDF overlay
\dontrun{
  d <- mixed_data("longitudinal")
  m <- mixed_bayes(height ~ time + (1 | subject), d, chains = 2, iter = 500)
  mixed_pp_check(m, "ecdf_overlay")
}
# Example 3: interval PPC
\dontrun{
  d <- mixed_data("longitudinal")
  m <- mixed_bayes(height ~ treatment * time + (1 | subject), d, chains = 2, 
      iter = 500)
  mixed_pp_check(m, "intervals")
}
}
