% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedFlowR-package.R
\docType{package}
\name{mixedFlowR-package}
\alias{mixedFlowR}
\alias{mixedFlowR-package}
\title{mixedFlowR: Design-Aware Mixed Models for Agronomic Experiments}
\description{
A design-aware orchestration layer for classical and modern mixed-model
analyses in agronomy and related experimental sciences. Estimation is
delegated to established R engines while the design, assumptions, inference
and provenance are retained in auditable result objects.
}
\author{
\strong{Maintainer}: Walter Esfrain Pereira \email{walterufpb@yahoo.com.br} (\href{https://orcid.org/0000-0003-1085-0191}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Walter Esfrain Pereira \email{walterufpb@yahoo.com.br} (\href{https://orcid.org/0000-0003-1085-0191}{ORCID}) [copyright holder]
  \item Magali Haidee Pereira Martinez (\href{https://orcid.org/0009-0009-5419-959X}{ORCID})
}

}
\keyword{internal}
