% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{splitplot_anova}
\alias{splitplot_anova}
\title{Extract classical split-plot ANOVA strata}
\usage{
splitplot_anova(object)
}
\arguments{
\item{object}{Result from \code{\link[=fit_legacy_splitplot]{fit_legacy_splitplot()}}.}
}
\value{
A tidy data frame of ANOVA strata.
}
\description{
Extract classical split-plot ANOVA strata
}
\examples{
d <- mixed_data("splitplot"); f <- fit_legacy_splitplot(d,"yield","block","variety","nitrogen")
# Example 1: standard extraction
splitplot_anova(f)
# Example 2: factor-coded nitrogen
d$nf <- factor(d$nitrogen)
splitplot_anova(fit_legacy_splitplot(d, "yield", "block", "variety", "nf"))
# Example 3: transformed response
d$logy <- log(d$yield)
splitplot_anova(fit_legacy_splitplot(d, "logy", "block", "variety", "nitrogen"))
}
